Rename omero_marshal.legacy.affinetransform module
