+++ omero_marshal/decode/decoders/affinetransform.py
@@ -17,7 +17,7 @@ try:
     from omero.model import AffineTransformI
 except ImportError:
     # Use internal AffineTransformI classes for OMERO 5.1.x and OMERO 5.2.x
-    from omero_marshal.legacy.AffineTransformI import AffineTransformI
+    from omero_marshal.legacy.affinetransform import AffineTransformI
 from omero.rtypes import rdouble
