+++ omero_marshal/legacy/affinetransform.py
